requirements.txt
# requirements.txt — Analytiq complete dependencies

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9   # PostgreSQL driver
alembic==1.13.1          # Database migrations

# Auth & Security
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0

# ML
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.45.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.0

# NLP / AI
groq==0.9.0
langchain==0.2.0

# Export
fpdf2==2.7.9
openpyxl==3.1.2

# Background jobs
celery==5.3.6
redis==5.0.4

# Monitoring / Logging
sentry-sdk[fastapi]==1.45.0

# HTTP
httpx==0.27.0
requests==2.31.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
